Rename girder.py to avoid relative import
