+++ plugins/oauth/server/providers/__init__.py
@@ -19,7 +19,7 @@
 
 import collections
 
-from .girder import Girder
+from .girder_provider import Girder
 from .google import Google
 from .globus import Globus
 from .github import GitHub
