+++ plugins/oauth/server/providers/girder_provider.py
